lab-04/aes-rsa-socket/requirements.txt
pycryptodome
PyQt5
requests